Ported code over to a better structure of folders, separated out modules
