+++ package.json
@@ -2,8 +2,8 @@
   "name": "pebblebed",
   "version": "0.5.1",
   "description": "Simplified interactions with Google Datastore for NodeJS",
-  "main": "dist/Pebblebed.js",
-  "typings": "dist/Pebblebed.d.ts",
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
   "engines": {
     "node": ">=6.5.0"
   },
