+++ src/index.ts
@@ -1,0 +1,8 @@
+import PebblebedModel from "./PebblebedModel";
+
+export * from "./Pebblebed";
+export * from "./types/PebblebedTypes";
+
+export {
+  PebblebedModel,
+}
